Improved LZSS search, -s option, & more!
* Windows standard streams are unfucked. Also, 'setvbuf' is called to
make sure the buffering mode is correct for the steam.
* LZSS search has been sped up a little, by rewriting the search routine
to use 'memchr' to do the search instead. Also, if a maximal length
match has been found the search is aborted, there is no point in looking
for other matches, which is not handled in the original code.
* An '-s' option has been added, for compressing strings and hex
dumping the compressed string.
* More documentation!
* A few more assertions have been added.
* The performance test should probably be removed, the removal of that
pathological case means the compression of all zeros is incredibly fast.
* 'unsigned' has replaced 'long'