Allow LZSS code to be more configurable
The LZSS code is more configurable, the minimum match length can now be
correctly raised and exploited, which is good for highly compressible
text consisting of many large runs of the same character or long
matches.

Some other code has been slightly reorganized so it can be made to be
more extensible and some new comments have been added.