Shared data-structure to reduce memory usage
A lot of small changes to effect this change, which is using a
shared data-structure to reduce memory usage. This development branch
may be removed, but it might be a good idea to implement an idea from
it, that is using a buffer passed into the "shrink" function as both the
initial dictionary contents and as a working memory, this would allow
greater user control.