Added user controllable buffer
A buffer that the user can control has been added, tests need to be
written to exercise all that this provides, as it allows a custom
initial dictionary, meaning small (think 100 byte) strings can be
compressed better, whilst not compromising on longer input, so long as
the input characteristics are known. It also allows the user to control
just where that input buffer is allocated.