Added Move-To-Front and Elias Gamma CODEC
Added two new CODECs, Elias Gamma coding, and a Move-To-Front encoder.
The MTF encoder can make data more compressible for the Elias Gamma
encoding, which is a universal code for encoding integers. It needs
documenting and better integrating into the project, it could perhaps be
used to encode LZSS length/distance pairs as well, which it might be
better suited to.